@@ -4,17 +4,17 @@ description: A PouchDB alike for Dart. Uses lawndart for local storage and Wilt 
 homepage: https://github.com/shamblett/sporran
 
 environment:
-  sdk: '>=2.0.0 <3.0.0'
+  sdk: '>=2.12.0-133.2.beta <3.0.0'
 
 dependencies:
   wilt: '^5.0.0'
-  json_object_lite: '^2.1.0'
+  json_object_lite: '^2.1.1'
 
 dev_dependencies:
-  test: '^1.14.2'
-  pana: '^0.14.3'
-  pedantic: '^1.9.0'
-  build_runner: '^1.10.0'
-  build_test: '^1.2.1'
-  build_web_compilers: '^2.8.0'
+  test: '^1.16.0-nullsafety.16'
+  pana: '^0.14.10'
+  pedantic: '^1.10.0-nullsafety.3'
+  build_runner: '^1.11.0'
+  build_test: '^1.3.4'
+  build_web_compilers: '^2.16.0'
 
